moved new game
