+++ main.py
@@ -14,70 +14,6 @@ from tinydb import TinyDB
 from packages.helper  import Helper
 import os
 
-class NeuesSpielDialog(QtWidgets.QDialog):
-    def __init__(self, kurse):
-        super().__init__()
-        self.setWindowTitle("Neues Spiel")
-        layout = QtWidgets.QVBoxLayout()
-        self.setFixedSize(400, 450)  # Set fixed size for the dialog window
-        self.setWhatsThis("Dieser Dialog ermöglicht es dir, ein neues Spiel zu starten. "
-                          "Wähle den Kurs und gib die Schlagzahlen ein.")
-
-        self.kurs_combo = QtWidgets.QComboBox()
-        self.kurs_combo.addItems(kurse)
-        layout.addWidget(QtWidgets.QLabel("Kurs:"))
-        layout.addWidget(self.kurs_combo)
-
-        # Neues Layout für die Schlagzahl-Eingabe in zwei Spalten
-        schlagzahl_layout = QtWidgets.QGridLayout()
-
-        self.schlagzahl_eingabe = {}
-        self.schlagzahl_labels = {}
-        for i in range(1, 19):
-            eingabe = QtWidgets.QLineEdit()
-            eingabe.setValidator(QtGui.QIntValidator())
-            self.schlagzahl_eingabe[i] = eingabe
-            label_text = f"Schlagzahl Loch {i}:"
-            label = QtWidgets.QLabel(label_text)
-            self.schlagzahl_labels[i] = label
-
-            # Anordnung in zwei Spalten
-            row = (i - 1) % 9  # Zeile 0-8
-            col = (i - 1) // 9  # Spalte 0 oder 1
-            schlagzahl_layout.addWidget(label, row, col * 2)  # Label in Spalte 0 oder 2
-            schlagzahl_layout.addWidget(eingabe, row, col * 2 + 1)  # Eingabe in Spalte 1 oder 3
-
-            if i > 9:
-                eingabe.hide()
-                label.hide()
-
-        layout.addLayout(schlagzahl_layout)  # Hinzufügen des GridLayout zum Hauptlayout
-
-        button_box = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        button_box.accepted.connect(self.accept)
-        button_box.rejected.connect(self.reject)
-        layout.addWidget(button_box)
-  
-        self.setLayout(layout)
-
-        self.kurs_combo.currentIndexChanged.connect(self.update_schlagzahl_eingabe)
-
-    def get_spiel_daten(self):
-        kurs = self.kurs_combo.currentText()
-        kurs_info = help.getCourseByID(kurs)
-        locher = len(kurs_info["par"])
-        schlagzahlen = []
-        for i in range(1, locher + 1):
-            try:
-                schlagzahlen.append(int(self.schlagzahl_eingabe[i].text()))
-            except ValueError:
-                schlagzahlen.append(0)
-
-        loch_bool = (locher == 9)
-        
-        help.addGame(kurs, schlagzahlen, loch_bool, 0, 0)
-        self.neun_loch_radio.toggled.connect(self.update_schlagzahl_eingabe)
-        self.achtzehn_loch_radio.toggled.connect(self.update_schlagzahl_eingabe)
 class NeuerKursDialog(QtWidgets.QDialog):
     def __init__(self):
         super().__init__()
@@ -331,10 +267,7 @@ class ExportScorecardDialog(QtWidgets.QDialog):
 
 class HandicapUI(QtWidgets.QWidget):
     def __init__(self):
-        super().__init__()
-        # create color palette
-        
-        
+        super().__init__()     
         # main component
         self.layout = QtWidgets.QVBoxLayout()
 
@@ -343,11 +276,9 @@ class HandicapUI(QtWidgets.QWidget):
         
 
         button_header = QtWidgets.QHBoxLayout()
-        self.neues_spiel_button = QtWidgets.QPushButton("Neues Spiel")
         self.neuer_kurs_button = QtWidgets.QPushButton("Neuer Kurs")
         self.kurs_löschen_button = QtWidgets.QPushButton("Kurs Löschen")
         self.export_scorecard_button = QtWidgets.QPushButton("Scorecard Exportieren")
-        button_header.addWidget(self.neues_spiel_button)
         button_header.addWidget(self.neuer_kurs_button)
         button_header.addWidget(self.kurs_löschen_button)
         button_header.addWidget(self.export_scorecard_button) 
@@ -375,7 +306,6 @@ class HandicapUI(QtWidgets.QWidget):
 
         self.setLayout(self.layout)
 
-        self.neues_spiel_button.clicked.connect(self.oeffne_neues_spiel_dialog)
         self.neuer_kurs_button.clicked.connect(self.neuer_kurs_hinzugefuegt)
         self.kurs_löschen_button.clicked.connect(self.kurs_loeschen_dialog)
         self.export_scorecard_button.clicked.connect(self.oeffne_export_scorecard_dialog) 
@@ -396,15 +326,6 @@ class HandicapUI(QtWidgets.QWidget):
 
         self.update()
 
-
-
-
-    def oeffne_neues_spiel_dialog(self):
-        kurse = help.getAllCourseIDs()
-        dialog = NeuesSpielDialog(kurse)
-        if dialog.exec_() == QtWidgets.QDialog.Accepted:
-            dialog.get_spiel_daten()
-            self.update()
 
     def neuer_kurs_hinzugefuegt(self):
         dialog = NeuerKursDialog()
