+++ qt_components/header_buttons.py
@@ -3,11 +3,14 @@ import matplotlib
 from packages.helper import Helper
 matplotlib.use('Qt5Agg')
 from PyQt5 import QtWidgets
+from qt_components.new_game_dialog import NewGameDialog
 
 
 class HeaderButton(QtWidgets.QWidget):
     def __init__(self, help: Helper, parent=None):
         super().__init__(parent)
+        self.help = help
+
         self.neues_spiel_button = QtWidgets.QPushButton("Neues Spiel")
         self.neuer_kurs_button = QtWidgets.QPushButton("Neuer Kurs")
         self.kurs_löschen_button = QtWidgets.QPushButton("Kurs Löschen")
@@ -26,3 +29,27 @@ class HeaderButton(QtWidgets.QWidget):
         self.export_scorecard_button.clicked.connect(self.oeffne_export_scorecard_dialog) 
 
         self.setLayout(layout)
+
+    def oeffne_neues_spiel_dialog(self):
+        kurse = self.help.getAllCourseIDs()
+        dialog = NewGameDialog(kurse, self.help)
+        if dialog.exec_() == QtWidgets.QDialog.Accepted:
+            dialog.get_spiel_daten()
+            self.parent().update()
+
+    def neuer_kurs_hinzugefuegt(self):
+        dialog = NewGameDialog()
+        if dialog.exec_() == QtWidgets.QDialog.Accepted:
+            dialog.get_kurs_daten()
+            self.parent().update()
+
+    def kurs_loeschen_dialog(self):
+        kurse = self.help.getAllCourseIDs()
+        dialog = KursLoeschenDialog(kurse)
+        dialog.exec_()
+        
+    def oeffne_export_scorecard_dialog(self):
+        dialog = ExportScorecardDialog()
+        if dialog.exec_() == QtWidgets.QDialog.Accepted:
+            kurs, is_whs, file_path, kurs_rating, slope_rating, stroke_indices, use_last_values = dialog.get_export_daten()
+            self.help.export_scorecard(kurs, is_whs, file_path, kurs_rating, slope_rating, stroke_indices, use_last_values)
